feat(forms-web-app/full-planning): refactor second stage
