+++ packages/forms-web-app/__tests__/unit/routes/full-planning/local-planning-department.test.js
@@ -1,17 +1,17 @@
-const { get, post } = require('../../../__tests__/unit/routes/router-mock');
-const localPlanningDepartmentController = require('../../controllers/full-planning/local-planning-department');
-const { validationErrorHandler } = require('../../validators/validation-error-handler');
-const fetchExistingAppealMiddleware = require('../../middleware/fetch-existing-appeal');
+const { get, post } = require('../router-mock');
+const localPlanningDepartmentController = require('../../../../src/controllers/full-planning/local-planning-department');
+const { validationErrorHandler } = require('../../../../src/validators/validation-error-handler');
+const fetchExistingAppealMiddleware = require('../../../../src/middleware/fetch-existing-appeal');
 const {
   rules: localPlanningDepartmentValidationRules,
-} = require('../../validators/full-planning/local-planning-department');
+} = require('../../../../src/validators/full-planning/local-planning-department');
 
-jest.mock('../../validators/full-planning/local-planning-department');
+jest.mock('../../../../src/validators/full-planning/local-planning-department');
 
 describe('routes/full-planning/local-planning-department', () => {
   beforeEach(() => {
     // eslint-disable-next-line global-require
-    require('./local-planning-department');
+    require('../../../../src/routes/full-planning/local-planning-department');
   });
 
   it('should define the expected routes', () => {
